+++ src/main/java/persistence/sql/ddl/constraint/Constraint.java
@@ -1,6 +1,6 @@
 package persistence.sql.ddl.constraint;
 
-import persistence.sql.ddl.utils.ColumnType;
+import persistence.sql.mapper.ColumnType;
 
 public interface Constraint {
