+++ src/main/java/persistence/sql/mapper/ColumnId.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.utils;
+package persistence.sql.mapper;
 
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
