+++ src/main/java/persistence/sql/mapper/ColumnType.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.utils;
+package persistence.sql.mapper;
 
 import persistence.sql.ddl.type.DataType;
