+++ src/main/java/persistence/sql/mapper/ColumnTypes.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.utils;
+package persistence.sql.mapper;
 
 import jakarta.persistence.Id;
