+++ src/main/java/persistence/sql/mapper/TableType.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl.utils;
+package persistence.sql.mapper;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
